requirements.txt
flask==3.0.0
transformers==4.40.0
tensorflow==2.20.0
sentencepiece>=0.1.99
numpy==1.26.0
scipy==1.11.4
deep-translator==1.11.4
python-dotenv==1.0.0
langdetect==1.0.9
requests==2.31.0
gunicorn
google-generativeai==0.3.2